dynabuffers-python/pyproject.toml
[tool.poetry]
name = "dynabuffers"
version = "0.18.0"
description = "Dynamic cross platform serialization library."
authors = ["Leftshift One <contact@leftshift.one>"]

[tool.poetry.dependencies]
python = "^3.5"
antlr4-python3-runtime = "4.7.2"

[tool.poetry.dev-dependencies]
pytest = "^3.0"
pytest-cov = "^2.8.1"
pytest-timeout = "^1.3"
toml = "^0.10.0"

[tool.poetry.scripts]
clean = "poetry_scripts:clean"
install = "poetry_scripts:install"
test = "poetry_scripts:test"
build = "poetry_scripts:build"
publish = "poetry_scripts:publish"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
